requirements.txt
python-telegram-bot
beautifulsoup4
aiohttp
python-dotenv
Telethon
curl_cffi